- Implement complete Apache Arrow object_store backend
- Achieve 15% better PUT performance vs native AWS SDK
- Add comprehensive performance testing framework
- Update documentation and version to 0.7.10
- Feature flag system for compile-time backend selection

Simple Performance Results:
- Arrow: 349.86 MB/s PUT, 2442.47 MB/s GET
- Native: 304.54 MB/s PUT, 2299.26 MB/s GET
- Arrow backend wins on all performance metrics